+++ cocoa-beans-commands/src/test/java/net/apartium/cocoabeans/commands/parsers/DurationParserTest.java
@@ -139,6 +139,23 @@ class DurationParserTest {
         assertParserTabCompletion(parser, null, null, args("2m"), 0, Set.of("2months", "2minutes", "2mills"), 1);
         assertParserTabCompletion(parser, null, null, args("2mo"), 0, Set.of("2months"), 1);
         assertParserTabCompletion(parser, null, null, args("2mi"), 0, Set.of("2minutes", "2mills"), 1);
+
+        assertParserTabCompletion(parser, null, null, args("2minutes 2"), 0, Set.of("20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "2months", "2minutes", "2mills", "2s"), 2);
+    }
+
+    @Test
+    void invalidTabCompletion() {
+        DurationParser parser = new DurationParser(DurationParser.DEFAULT_KEYWORD, 0, Map.of(
+                "months", Duration.ofDays(30),
+                "minutes", Duration.ofMinutes(1),
+                "mills", Duration.ofMillis(1),
+                "s", Duration.ofSeconds(1)
+        ));
+
+        assertParserTabCompletion(parser, null, null, args("2minutes 2a"), 0, null, -1);
+        assertParserTabCompletion(parser, null, null, args("wow"), 0, null, -1);
+        assertParserTabCompletion(parser, null, null, args("2147483647a"), 0, null, -1);
+        assertParserTabCompletion(parser, null, null, args("213a4"), 0, null, -1);
     }
 
 }